introduce FlatEmbed, remove flat_resolver, make sure calls to FlatEmbed are resolved in flattening, make flattening inlining stricter
